Release 1.1.0
Parse swupdate.cfg to get version info for UI.

Add code to handle ustate variable for updates.  When SWUpdate updates
the system from a local source (i.e USB, mongoose, etc.) the ustate
variable is not changed...this is left to the app or script to handle.
In the case of hawkbit, ustate is handled automatically.

Signed-off-by: Matt Wood <matt.wood@microchip.com>